+++ encoders/log_util.py
@@ -5,7 +5,7 @@ from opyenxes.classification.XEventAttributeClassifier import XEventAttributeCla
 TIME_FORMAT = "%Y-%m-%dT%H:%M:%S"
 DEFAULT_COLUMNS = ["trace_id", "event_nr", "remaining_time", "elapsed_time"]
 TIMESTAMP_CLASSIFIER = XEventAttributeClassifier("Timestamp", ["time:timestamp"])
-
+HEADER_COLUMNS = ['trace_id', 'remaining_time', 'elapsed_time']
 
 def unique_events(log: list):
     """List of unique events using event concept:name
